requirements.txt
djangorestframework==3.14.0
Django==4.0.5
djangorestframework-simplejwt==5.3.1
